# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для платформы для мастеров ручной работы",
        "title": "CraftPlace",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/posts": {
            "get": {
                "description": "Возвращает список постов с возможностью фильтрации по магазину",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Поиск"
                ],
                "summary": "Получить посты",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Фильтр по ID магазина",
                        "name": "id_shop",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список постов",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/reqresp.PostResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Неверный формат ID магазина",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/products": {
            "get": {
                "description": "Возвращает список товаров с возможностью фильтрации по различным параметрам",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Поиск"
                ],
                "summary": "Получить товары",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Фильтр по названию товара",
                        "name": "title",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Минимальная цена товара",
                        "name": "min_cost",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 100000,
                        "description": "Максимальная цена товара",
                        "name": "max_cost",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "uuid",
                        "default": "00000000-0000-0000-0000-000000000000",
                        "description": "Фильтр по ID магазина",
                        "name": "id_shop",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "uuid",
                        "default": "00000000-0000-0000-0000-000000000000",
                        "description": "Фильтр по ID категории",
                        "name": "id_category",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список товаров",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/reqresp.ProductResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Неверный формат параметров",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/user/user-posts": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Добавляет новый пост в указанный магазин пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Посты"
                ],
                "summary": "Добавить пост в магазин",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer токен",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Данные нового поста",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/reqresp.AddPostRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Пост успешно добавлен",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Удаляет пост пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Посты"
                ],
                "summary": "Удалить пост",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer токен",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Данные для удаления поста",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/reqresp.DeletePostRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Пост успешно удален",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/user/user-products": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Обновляет данные товара пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Изделия"
                ],
                "summary": "Обновить товар",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer токен",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Данные для обновления товара",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/reqresp.UpdateProductRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Товар успешно обновлен",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Добавляет новый товар в указанный магазин пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Изделия"
                ],
                "summary": "Добавить товар в магазин",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer токен",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Данные нового товара",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/reqresp.AddProductRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Товар успешно добавлен",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Удаляет товар пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Изделия"
                ],
                "summary": "Удалить товар",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer токен",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Данные для удаления товара",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/reqresp.DeleteProductRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Товар успешно удален",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "reqresp.AddPostRequest": {
            "type": "object",
            "required": [
                "description",
                "shopID"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "Магазин сережек"
                },
                "shopID": {
                    "type": "string",
                    "example": "bb2e8400-e29b-41d4-a716-446655442222"
                }
            }
        },
        "reqresp.AddProductRequest": {
            "type": "object",
            "required": [
                "categoryIDs",
                "cost",
                "description",
                "shopID",
                "title"
            ],
            "properties": {
                "categoryIDs": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "cost": {
                    "type": "integer",
                    "minimum": 0,
                    "example": 100
                },
                "description": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "Магазин сережек"
                },
                "shopID": {
                    "type": "string",
                    "example": "bb2e8400-e29b-41d4-a716-446655442222"
                },
                "title": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "Звезды"
                }
            }
        },
...
        "reqresp.DeletePostRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "bb2e8400-e29b-41d4-a716-446655442222"
                }
            }
        },
        "reqresp.DeleteProductRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "bb2e8400-e29b-41d4-a716-446655442222"
                }
            }
        },
...
        "reqresp.PostResponse": {
            "type": "object",
            "required": [
                "description",
                "shopID"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "Лучший магазин сережек"
                },
                "id": {
                    "type": "string",
                    "example": "bb2e8400-e29b-41d4-a716-446655442222"
                },
                "shopID": {
                    "type": "string",
                    "example": "bb2e8400-e29b-41d4-a716-446655442222"
                },
                "timePublication": {
                    "type": "string",
                    "example": "2023-06-15T10:00:00Z"
                }
            }
        },
        "reqresp.ProductResponse": {
            "type": "object",
            "required": [
                "categoryIDs",
                "cost",
                "description",
                "shopID"
            ],
            "properties": {
                "categoryIDs": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "cost": {
                    "type": "integer",
                    "minimum": 0,
                    "example": 200
                },
                "description": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "Лучший магазин сережек"
                },
                "id": {
                    "type": "string",
                    "example": "bb2e8400-e29b-41d4-a716-446655442222"
                },
                "shopID": {
                    "type": "string",
                    "example": "bb2e8400-e29b-41d4-a716-446655442222"
                },
                "title": {
                    "type": "string",
                    "example": "Eco"
                }
            }
        },
...
        "reqresp.UpdateProductRequest": {
            "type": "object",
            "required": [
                "categoryIDs",
                "cost",
                "description",
                "shopID",
                "title"
            ],
            "properties": {
                "categoryIDs": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "cost": {
                    "type": "integer",
                    "minimum": 0,
                    "example": 200
                },
                "description": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "Лучший магазин сережек"
                },
                "id": {
                    "type": "string",
                    "example": "bb2e8400-e29b-41d4-a716-446655442222"
                },
                "shopID": {
                    "type": "string",
                    "example": "bb2e8400-e29b-41d4-a716-446655442222"
                },
                "title": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "Лучшие звезды"
                }
            }
        },
...
    }
}